REFACTOR mapping logic in seperate context
